backend/requirements.txt
fastapi
uvicorn
joblib
pandas
scikit-learn
python-dotenv